requirements.txt
# Core ML dependencies for HTR
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Image processing
opencv-python>=4.7.0
scikit-image>=0.20.0
Pillow>=9.5.0
numpy>=1.24.0

# Text processing and NLP
pyspellchecker>=0.7.0
# language-tool-python>=2.7.0  # Optional: uncomment for grammar checking (requires Java)

# Data handling
datasets>=2.12.0
pandas>=2.0.0

# Evaluation metrics
evaluate>=0.4.0
jiwer>=3.0.0  # For WER/CER calculation

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
PyYAML>=6.0

# Web interface (optional)
streamlit>=1.31.0
sentence-transformers>=2.2.0
gTTS>=2.5.0

# Legacy support (optional)
# ollama>=0.1.0
# anthropic>=0.49.0
